[app/d_angulo] build method implemented in app.py

+++ app/d_angulo/app.py
@@ -1,5 +1,5 @@
 """
-Angulo
+Angulo web app
 """
 
 from lib import base_app, build, http, image
@@ -31,17 +31,46 @@ class app(base_app):
 	"""
         program build/update
         """
-        # Create bin dir (delete the previous one if exists)
-        if os.path.isdir(self.bin_dir):
-            shutil.rmtree(self.bin_dir)
-        os.mkdir(self.bin_dir)
+        # store common file path in variables
+        tgz_file = self.dl_dir + "all.tgz"
+        prog_file = self.bin_dir + "stereoSSD-mean"
+        log_file = self.base_dir + "build.log"
 
+        # get the latest source code
+	# TODO : download only if the date of the latest source archive is newer
+        build.download("http://dev.ipol.im/git/?p=facciolo/stereo.git;a=snapshot;h=HEAD;sf=tgz",
+			 tgz_file)
+
+        # test if one of the binaries ("prog_file") is missing, or too old
+        if (os.path.isfile(prog_file)
+            and ctime(tgz_file) < ctime(prog_file)):
+            cherrypy.log("not rebuild needed", context='BUILD', traceback=False)
+        else:
+            # Create bin dir (delete the previous one if exists)
+            if os.path.isdir(self.bin_dir):
+                shutil.rmtree(self.bin_dir)
+            os.mkdir(self.bin_dir)
+
+            # extract the archive
+            build.extract(tgz_file, self.src_dir)
+            # ammend the source
+            onlypath = os.listdir(self.src_dir)[0]
+            print ( str(onlypath) )
+            self.src_dir = self.src_dir+'/'+onlypath;
+            # build the program
+            build.run("make -C %s %s" % (self.src_dir, "all"), stdout=log_file)
+
+            # cleanup the source dir
+            shutil.rmtree(self.src_dir)
+            os.remove(tgz_file)
+	    
 	# link all the scripts to the bin dir
         import glob
         for file in glob.glob( os.path.join( self.base_dir, 'scripts/*')):
             os.symlink(file, os.path.join( self.bin_dir , os.path.basename(file)))
-
+	
 	return
+
 
     def __init__(self):
         """
